Fixed Ultra Ball oversight
Previously, Great Balls were the only Pokeballs to have a ballfactor of 8, while the rest were 12.

A lower ballfactor leads to a higher catch rate, therefore it seems to be an oversight that the
Great Ball can actually be better than the Ultra Ball in some cases.

Now, Ultra Balls and Great Balls both have a ballfactor of 8, allowing for the smaller random range of
and Ultra Ball to always be superior to a Great Ball.